Migrate privilege test to sql logic tester
Move TestInsecure back to driver_test
